# Optimize SpeedTestClass::update()